Refactor CreateInternshipInterface. Break larger componenets out into their own files.
